Move editor placeholder into iframe
