+++ src/components/editor/prosemirror.js
@@ -1,11 +1,13 @@
-import { forwardRef, useEffect, useImperativeHandle, useState } from 'react'
+import { forwardRef, useEffect, useImperativeHandle, useMemo, useState } from 'react'
+import { useIntl } from 'react-intl'
 import { bool, func, instanceOf, string } from 'prop-types'
 import { EditorState } from 'prosemirror-state'
 import { EditorView } from 'prosemirror-view'
 import cx from 'classnames'
 import { Frame } from '../frame.js'
-import { nodeViews } from '../../editor/schema.js'
+import { isBlank, nodeViews } from '../../editor/index.js'
 import { useEvent } from '../../hooks/use-event.js'
+import { create } from '../../dom.js'
 import { isMeta } from '../../keymap.js'
 
 
@@ -19,10 +21,16 @@ export const ProseMirror = forwardRef(({
   onContextMenu,
   onFocus,
   onKeyDown,
+  placeholder,
   state,
   wrap
 }, ref) => {
   let [view, setView] = useState()
+  let intl = useIntl()
+
+  let p = useMemo(() => (
+    create('div', { class: 'placeholder' })
+  ), [])
 
   let editable = useEvent(() =>
     !(isDisabled || isReadOnly))
@@ -45,6 +53,8 @@ export const ProseMirror = forwardRef(({
       nodeViews,
       state
     }))
+
+    doc.body.prepend(p)
   })
 
   let handleUnload = useEvent(() => {
@@ -56,10 +66,20 @@ export const ProseMirror = forwardRef(({
     view?.updateState(state)
   }, [state, view])
 
+
   useImperativeHandle(ref, () => view, [view])
 
-  // TODO handle container click (links)
+  useEffect(() => {
+    p.textContent = intl.formatMessage({ id: placeholder })
+  }, [placeholder, intl, p])
 
+  useEffect(() => {
+    p.style.display = (
+      isDisabled || isReadOnly || !isBlank(state.doc)
+    ) ? 'none' : 'block'
+  }, [state, isDisabled, isReadOnly, p])
+
+  // TODO handle container click (links)
 
   return (
     <Frame
@@ -75,6 +95,8 @@ export const ProseMirror = forwardRef(({
 ProseMirror.propTypes = {
   isDisabled: bool,
   isReadOnly: bool,
+  mode: string,
+  numbers: bool,
   // Subtle: event handlers are passed to PM on initialization
   // and they will not be updated. Use stable references!
   onBlur: func,
@@ -82,11 +104,9 @@ ProseMirror.propTypes = {
   onContextMenu: func,
   onFocus: func,
   onKeyDown: func.isRequired,
-  mode: string,
-  numbers: bool,
-  wrap: bool,
-
-  state: instanceOf(EditorState)
+  placeholder: string,
+  state: instanceOf(EditorState),
+  wrap: bool
 }
 
 ProseMirror.propTypes = {
